+++ composer.json
@@ -6,7 +6,7 @@
     "license": "MIT",
     "require": {
         "lab404/laravel-impersonate": "^1.7",
-        "filament/filament": "^3.0"
+        "filament/filament": "^4.0"
     },
     "autoload": {
         "psr-4": {
